Separated web app into domains
