+++ web/marketplaces_web.py
@@ -1,0 +1,42 @@
+from flask import Blueprint, render_template, request, redirect, url_for
+
+from backend.controllers.marketplace_controller import MarketplaceController
+from backend.models.marketplace import Marketplace
+
+marketplaces = Blueprint('marketplace', __name__)
+
+marketplace_controller = MarketplaceController()
+
+
+@marketplaces.route('/marketplaces', methods=['GET', 'POST'])
+def list_marketplace():
+    if request.method == "POST":
+        name = request.form.get('name')
+        description = request.form.get('description')
+        marketplace = Marketplace(name, description)
+        marketplace_controller.create(marketplace)
+    marketplaces = marketplace_controller.read_all()
+    return render_template('list_marketplace.html', title='Marketplaces', list=marketplaces)
+
+
+@marketplaces.route('/create_marketplace')
+def new_marketplace():
+    return render_template('create_marketplace.html', title='Novo Marketplace')
+
+
+@marketplaces.route('/marketplaces/<int:id>', methods=['GET', 'POST'])
+def edit_marketplace(id: int):
+    if request.method == "POST":
+        name = request.form.get('name')
+        desc = request.form.get('description')
+        new_marketplace = Marketplace(name, desc, id)
+        marketplace_controller.update(new_marketplace)
+        return redirect(url_for('list_marketplace'))
+    marketplace = marketplace_controller.read_by_id(id)
+    return render_template('edit_marketplace.html', title='Marketplace', object=marketplace)
+
+
+@marketplaces.route('/marketplaces/<int:id>/delete', methods=['GET'])
+def erase_marketplace(id: int):
+    marketplace_controller.delete(id)
+    return redirect(url_for('list_marketplace'))
